chore(algorithms): test files, parsers and algorithms

+++ src/problem/Customer.hpp
@@ -1,21 +1,22 @@
 #pragma once
 #include <vector>
+
 /**
-* @brief Class representing a customer in a UFLP.
-*/
+ * @brief Class representing a customer in a UFLP.
+ */
 class Customer {
 private:
-	int demand;
-	std::vector<int> allocation_costs;
+    int demand;
+    std::vector<double> allocation_costs;
 
 public:
-	Customer(int dem, const std::vector<int>& costs) : demand(dem), allocation_costs(costs) {}
+    Customer(int dem, const std::vector<double>& costs) : demand(dem), allocation_costs(costs) {}
 
-	int getDemand() const {
-		return demand;
-	}
+    int getDemand() const {
+        return demand;
+    }
 
-	const std::vector<int>& getAllocationCosts() const {
-		return allocation_costs;
-	}
+    const std::vector<double>& getAllocationCosts() const {
+        return allocation_costs;
+    }
 };
